requirements.txt
langchain-classic
langchain
langgraph
langchain-openai
langchain-huggingface
langchain-groq
langchain-core
langchain-ollama

openai

langgraph-supervisor

langchain-chroma
langchain-text-splitters
langchain-community
chromadb
sentence-transformers
pypdf
python-docx

duckduckgo-search
ddgs
icalendar

streamlit

python-dotenv
pydantic
uvicorn
fastapi

boto3